in bin

+++ controllers/passwordsController.js
@@ -42,7 +42,14 @@ const deletePassword = async (req, res) => {
     if (!password) {
         return res.status(204).json({ "message": `No password matches ID ${req.body.id}.` });
     }
-    const result = await password.deleteOne(); //{ _id: req.body.id }
+    let result = {}
+    if(password.in_bin)
+        result = await password.deleteOne(); //{ _id: req.body.id }
+    else
+        result = await Password.findOneAndUpdate((query, {title:password.title,password:password.password,autor:password.autor,in_bin:true}, {upsert: true}, function(err, doc) {
+            if (err) return res.send(500, {error: err});
+            return res.send('Succesfully saved.');
+        }))
     res.json(result);
 }
